feat(order): using unit of work

+++ ShopDBProduct/Entities/Order.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public int Price { get; set; } = 0;
+        public decimal Price { get; set; } = 0;
         public ICollection<OrderItem> Items { get; set; } = new List<OrderItem>();
     }
 }
